requirements.txt
scikit-learn
pandas
psycopg2-binary
joblib
requests